Use Live Movie Data - More Info Button
The id parameter in the Route's path will be bound to the id of the clicked movie, e.g. 475557. The Route causes the app to mount the Movie Details page. That page's logic must request the movie's details from the TMDB API - previously, we passed in the details as a prop, but this was a temporary solution. In src/pages/movieDetailsPage.js, make the these modifications: