The Filtering Behaviour -
A useState hook to manage the state variables.
A function to update the state.
Logic to compute the subset of matching movies.
A 'floating action' button.

The FilterMoviesCard component manages the filtering input fields. However, the HomePage component uses these values to compute the matching list of movies. The design solution for this problem is the Inverse data flow pattern or Data Down, Action Up. According to the pattern, the HomePage component stores the filtering data as 'state'.