Tagging Favourite Movies - Add to favourites logic completed; Dummy favourites page.
The app has a bug! Try the following sequence:

Go to the home page.
Click the 'Favourites' icon of the first three movies.
Notice, the selected movies are tagged as expected.
Go to the Favourites page, using the navigation links on top
Navigate back to the Home page. The three selected movies are no longer tagged. This reset happened because the useEffect hook queries the API every time the home page component is mounted, i.e. we overwrite the tagged movies information. We will fix this issue in the next lab